Emergent Meeting Minutes
Added notes for our emergent meeting. We discussed the issues with
FTB, what happened, what is at risk, and our proposed next steps.
This was disclosed to all members who were at the space in person
and we plan to disclose all of this info at the next member meeting
in two weeks.